# server/openapi/v1.yaml
openapi: 3.0.0
info:
  title: Immigration Horizons Staff API
  version: "1.0"
  description: Canonical JSON API for the Angular Enterprise UI.
servers:
  - url: /api/v1
paths:
...
  /staff/cases/{caseId}/stage:
    patch:
      summary: Advance or update case stage
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stage]
              properties:
                stage: { type: string }
      responses:
        '200':
          description: Updated case details
        '403':
          description: Permission denied
  /staff/cases/{caseId}/project-manager:
    patch:
      summary: Assign or change Project Manager for case
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId: { type: string, nullable: true }
      responses:
        '200':
          description: Updated case details
  /staff/cases/{caseId}/members:
    post:
      summary: Add employee team member to case workspace
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeId]
              properties:
                employeeId: { type: string }
                role: { type: string, default: contributor }
      responses:
        '200':
          description: Updated case details
  /staff/cases/{caseId}/members/{memberId}:
    delete:
      summary: Remove member from case workspace
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
        - name: memberId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated case details
  /staff/cases/{caseId}/archive:
    post:
      summary: Archive case (non-destructive)
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated case details
  /staff/cases/{caseId}/client-updates:
    post:
      summary: Publish client update message for case portal
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, maxLength: 1000 }
      responses:
        '201':
          description: Published update message document
...